Replaced Mautic form with HubSpot form
We are migrating all of our marketing automation assets from Mautic to HubSpot. I replaced the Mautic form with the HubSpot version. We might need to check and see how this form looks on the site; I set the text color to white, but we might need to tweak it a little to get it to look good.