Created a mailing list subscription form
This HubSpot form can be used once we migrate the blog over from the OCI website. No sales person will contact users who fill out this form; they'll just be added to the 2GM mailing list. I made the text in this form white, but if the background on the blog pages doesn't end up being black, we can change it.