Replaced Mautic form with HubSpot form
We are migrating all of our marketing automation from Mautic over to HubSpot. This form should work the same way the old one did. I made the text and the button white to show up on a black back ground; if the page design changes, I will update the styles.