Change binary download link
Change binary download link to point to micronaut-starter releases page instead of the micronaut-core source zip.

It's possible the original documentation meant to download the source, but newer starter documentation implies that the download will include the mn executable. If team still wants to provide a link to download the source package as a zip here, perhaps it should be two different links. 

Resolves #219